requirements.txt
Flask
Gunicorn
openai==0.27.4
python-dotenv==1.0.0
langchain
chromadb
unstructured==0.7.12
